Electrical hourly update
New query to calculate hourly electrical energy usage.
The query calculates the average W per minute and then determines how many minutes in each hour have data. The energy is proportional to that ratio.
For the running hour the ratio is relative to the amount of minutes passed instead of the whole hour.